refactor(docs): replace react ui stack component
