+++ packages/apps/docs/src/components/NotFound/styles.ts
@@ -1,4 +1,5 @@
-import { Stack, styled, StyledComponent } from '@kadena/react-components';
+import { styled, StyledComponent } from '@kadena/react-components';
+import { Stack } from '@kadena/react-ui';
 
 export const Section: StyledComponent<typeof Stack> = styled(Stack, {
   padding: '$20 0 0',
