+++ packages/apps/docs/src/pages/docs/blogchain/2018/index.tsx
@@ -1,4 +1,5 @@
-import { Heading, Stack } from '@kadena/react-components';
+import { Heading } from '@kadena/react-components';
+import { Stack } from '@kadena/react-ui';
 
 import {
   checkSubTreeForActive,
@@ -9,7 +10,7 @@ import React, { FC } from 'react';
 
 const Home: FC = () => {
   return (
-    <Stack direction="column" spacing="2xl">
+    <Stack direction="column" spacing="$2xl">
       <div>
         <Heading as="h2">BlogChain 2018</Heading>
       </div>
