+++ packages/apps/docs/src/pages/docs/chainweb/index.tsx
@@ -1,5 +1,4 @@
-import { Stack } from '@kadena/react-components';
-import { Heading } from '@kadena/react-ui';
+import { Heading, Stack } from '@kadena/react-ui';
 
 import {
   checkSubTreeForActive,
@@ -10,7 +9,7 @@ import React, { FC } from 'react';
 
 const Home: FC = () => {
   return (
-    <Stack direction="column" spacing="2xl">
+    <Stack direction="column" spacing="$2xl">
       <div>
         <Heading as="h2">Chainweb</Heading>
       </div>
