+++ packages/apps/docs/src/pages/docs/contribute/index.tsx
@@ -1,4 +1,4 @@
-import { Stack } from '@kadena/react-components';
+import { Stack } from '@kadena/react-ui';
 
 import { BrowseSection } from '@/components';
 import {
@@ -10,7 +10,7 @@ import React, { FC } from 'react';
 
 const Home: FC = () => {
   return (
-    <Stack direction="column" spacing="2xl">
+    <Stack direction="column" spacing="$2xl">
       <BrowseSection>
         <BrowseSection.LinkBlock
           title="Run a Node"
