+++ packages/apps/docs/src/pages/docs/kadena/index.tsx
@@ -1,5 +1,4 @@
-import { Stack } from '@kadena/react-components';
-import { Heading, Text } from '@kadena/react-ui';
+import { Heading, Stack, Text } from '@kadena/react-ui';
 
 import { browseSectionWrapper, fullWidth } from '../../../styles/index.css';
 
@@ -14,7 +13,7 @@ import React, { FC } from 'react';
 
 const Home: FC = () => {
   return (
-    <Stack direction="column" spacing="2xl">
+    <Stack direction="column" spacing="$2xl">
       <div>
         <Heading as="h2">Welcome to Kadena&apos;s documentation!</Heading>
         <Text>
@@ -40,7 +39,7 @@ const Home: FC = () => {
         </Text>
       </div>
 
-      <Stack flexWrap="wrap" spacing="2xs" className={fullWidth}>
+      <Stack wrap="wrap" spacing="$2xs">
         <BrowseSection title="General" className={browseSectionWrapper}>
           <Link href="/docs/kadena/overview">Overview of Kadena</Link>
           <Link href="/docs/kadena/whitepapers">Whitepapers</Link>
